Allow deploy to CF pages

+++ deploy.sh
@@ -16,23 +16,31 @@ jq --raw-output -c '.[]' sites.json | while read i; do
 
   export SITE="$site"
   deploy_path="$(echo "$i" | jq --raw-output -c '.deployPath')"
+  pages="$(echo "$i" | jq --raw-output -c '.pages')"
 
-  if [ "$HOST" = "localhost" ] || [ "$HOST" = "127.0.0.1" ]; then
-    mkdir -p "$deploy_path"
-    dest="$deploy_path"
-  elif [ -z "$HOST" ]; then
-    echo -n "HOST environment variable is not set. Please set it to "
-    echo -n "the remote host you want to deploy to or to localhost "
-    echo "to have it deployed in this machine."
-    exit 1
-  else
-    ssh "$HOST" "mkdir -p $deploy_path"
-    dest="$HOST:$deploy_path"
+  if [ -z "$pages" ]; then
+    if [ "$HOST" = "localhost" ] || [ "$HOST" = "127.0.0.1" ]; then
+      mkdir -p "$deploy_path"
+      dest="$deploy_path"
+    elif [ -z "$HOST" ]; then
+      echo -n "HOST environment variable is not set. Please set it to "
+      echo -n "the remote host you want to deploy to or to localhost "
+      echo "to have it deployed in this machine."
+      exit 1
+    else
+      ssh "$HOST" "mkdir -p $deploy_path"
+      dest="$HOST:$deploy_path"
+    fi
   fi
 
   echo =========================================================
   echo "Deploying $site"
   bun run scripts/fetch-data.js -s "$site"
   bun run build
-  rsync -avz --delete dist/ "$dest"
+
+  if [ -n "$pages" ]; then
+    wrangler pages deploy dist --project-name "$site"
+  else
+    rsync -avz --delete dist/ "$dest"
+  fi
 done
